requirements.txt
# ShadowNet Nexus - Gemini-Powered Anti-Forensics Detection
# Core Dependencies

# Gemini API
google-generativeai>=0.3.0

# System Monitoring
psutil>=5.9.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Data Processing
pillow>=10.0.0

# Data Validation & Security
pydantic>=2.0.0

# Utilities
requests>=2.31.0

# API Server (for Dashboard)
flask>=3.0.0
flask-cors>=4.0.0

# Optional: Enhanced Features
# colorama>=0.4.6  # Colored terminal output
# rich>=13.0.0     # Rich terminal formatting
